move build output to single directory
