+++ .gitignore
@@ -1,3 +1,2 @@
-artifacts/
-cache/
+build/
 node_modules/
