+++ buidler.config.ts
@@ -1,9 +1,11 @@
-import { task, usePlugin } from "@nomiclabs/buidler/config";
+import { usePlugin } from "@nomiclabs/buidler/config";
 
 usePlugin("@nomiclabs/buidler-waffle");
 
 export default {
   paths: {
+    artifacts: "build/artifacts",
+    cache: "build/cache",
     sources: "src/contracts",
   },
   solc: {
